代码生成器增加同步数据结构； 修复sql server 生成float类型; 修复非system用户【代码生成器中找不到xxx的定义】； 修复用户已处理流程列表中看不到【不同意】的流程；

+++ OpenAuth.App/DbExtension.cs
@@ -278,7 +278,7 @@ select utc.column_name as                           COLUMNNAME,
                      THEN 'DateTime'
                      WHEN ColumnType IN('smallmoney', 'DECIMAL', 'numeric',
                                           'money') THEN 'decimal'
-                     WHEN ColumnType = 'float' THEN 'float'
+                     WHEN ColumnType = 'float' THEN 'double'
                      ELSE 'string'
                 END as  EntityType,
                     ColumnType,
